Inline on create comment trigger into api call
